requirements.txt
# requirements.txt
# pip install -r requirements.txt
flask
# matplotlib
pdfminer.six
pandas